+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Added additional window decoration.
 - Added footer help mentioning keyboard shortcuts.
+- Added a couple of default commands if there is no command database.
 - Fixed wasted space when not displaying window decoration.
 - Fixed auto run when search term has spaces.